old/requirements.txt

google-genai>=0.2.0
gradio>=4.0.0
Pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.24.0
